refactor: domain 하위의 repository 패키지 생성
