--- src/main/java/com/backend/connectable/event/domain/EventRepositoryCustom.java
@@ -1,4 +1,0 @@
-package com.backend.connectable.event.domain;
-
-public interface EventRepositoryCustom {
-}
