+++ src/main/java/com/backend/connectable/event/domain/repository/EventRepository.java
@@ -1,5 +1,6 @@
-package com.backend.connectable.event.domain;
+package com.backend.connectable.event.domain.repository;
 
+import com.backend.connectable.event.domain.Event;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface EventRepository extends JpaRepository<Event, Long>, EventRepositoryCustom {
