+++ src/main/java/com/backend/connectable/event/domain/repository/EventRepositoryCustom.java
@@ -1,0 +1,4 @@
+package com.backend.connectable.event.domain.repository;
+
+public interface EventRepositoryCustom {
+}
