+++ src/main/java/com/backend/connectable/event/domain/repository/EventRepositoryImpl.java
@@ -1,4 +1,4 @@
-package com.backend.connectable.event.domain;
+package com.backend.connectable.event.domain.repository;
 
 import org.springframework.stereotype.Repository;
