+++ src/main/java/com/backend/connectable/event/service/EventService.java
@@ -1,6 +1,6 @@
 package com.backend.connectable.event.service;
 
-import com.backend.connectable.event.domain.EventRepository;
+import com.backend.connectable.event.domain.repository.EventRepository;
 import com.backend.connectable.event.ui.dto.EventResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
