+++ src/main/java/com/backend/connectable/user/domain/repository/UserRepository.java
@@ -1,6 +1,7 @@
-package com.backend.connectable.user.domain;
+package com.backend.connectable.user.domain.repository;
 
 
+import com.backend.connectable.user.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
