+++ src/test/java/com/backend/connectable/user/domain/UserRepositoryTest.java
@@ -1,5 +1,6 @@
 package com.backend.connectable.user.domain;
 
+import com.backend.connectable.user.domain.repository.UserRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
